refactor: move state generics to entities layer
