+++ src/entities/state/index.ts
@@ -1,0 +1,4 @@
+export { default as State } from "./model/StatePublisher";
+
+// TODO: get rid of this state altogether, and use StatePublisher instead
+export { default as StateAuth } from "./model/State";
