+++ src/entities/state/model/State.ts
