+++ src/features/game/model/HintSettings.ts
@@ -1,4 +1,4 @@
-import State from "../../../app/state/StatePublisher";
+import { State } from "../../../entities/state";
 
 export interface HintSettingsData {
   translation: boolean;
